More advanced and graph-like graphql api

+++ src/main/kotlin/money/tegro/market/contract/RoyaltyContract.kt
@@ -18,7 +18,7 @@ data class RoyaltyContract(
     @JsonSerialize(using = MsgAddressSerializer::class)
     val destination: MsgAddress,
 ) {
-    fun value() = numerator.toFloat() / denominator
+    fun value() = numerator.toDouble() / denominator
 
     fun createData() = CellBuilder.createCell {
         storeUInt(numerator, 16)
